refactor Nodes and NodeSelection
